Move ZTP Left shifting job to a public repo location (#58608)
* Move ZTP Left shifting job to a public repo location

Signed-off-by: Carlos Cardenosa <ccardeno@redhat.com>

* Fixed incorrect format

Signed-off-by: Carlos Cardenosa <ccardeno@redhat.com>

* Increase Prow default timeout

The intention of this commit is to allow that,
in the event that the installation is not completed
in the maximum estimated time, the job ends in an
orderly manner, cleaning up the components created
ex-process for the spoke deployment.

Signed-off-by: Carlos Cardenosa <ccardeno@redhat.com>

* Install and use LVM Operator

Due to some issue related to NFS and assisted installer
a new Storage Class will be deployed via LVM Operator
to be used during the Spoke cluster deployment.

Signed-off-by: Carlos Cardenosa <ccardeno@redhat.com>

* Troubleshooting info

All SSH keys are generated during the current deployment
(therefore, ephemeral) and so are valid only for each run.

In case of having to look into the node being installed
its SSH private key is needed. Additionally, to get access
to this SSH key info access to a private network is required.

Therefore, it is safe to do this and have a way to troubleshoot
potential issues during the normal job operation.

Dump this info into the logs are useless, as it ends up
being removed by Prow internal processes.

Signed-off-by: Carlos Cardenosa <ccardeno@redhat.com>

---------

Signed-off-by: Carlos Cardenosa <ccardeno@redhat.com>